Updated Wildcard.py to fix my previous pull request on Wildcard-Multi-Select-Quantifier. (#586)
* Update wildcards.py

Fixed a big issue of the new wildcard multi-select-pattern quantifier that replaced all sub-subsequent occurrences of the wildcard keyword with the 1st occurrence's keyword, utterly destroying the prompt's identity.

* Update wildcards.py

Further fix for when using "wildcard" character '*' in __wildcard/*__ in combination wildcard-multi-select-pattern-quantifier.  Previous patch breaks the "wildcard" in wildcard scenario and wouldn't work properly.