Update core.py
Optimize two pieces of code, under the enhance_detail_for_animatediff function, use torch.stack instead of torch.cat for the upscale process of noise_mask, and use torch.cat instead of the displayed for loop for the latent_frames processing. For the latent_frames process, use torch.cat instead of the display for loop.